// src/lib/api/types.ts
export interface operations {
    getApiV1Job: {
        parameters: {
            query?: {
                page?: number;
                pageSize?: number;
                workspaceId?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: number;
                        title: string;
                        url: string | null;
                        description: string | null;
                        status: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                        nextEventDate: Record<string, never> | null;
                        lastCommuncationDate: Record<string, never> | null;
                        notes: string | null;
                        aiSummary: string | null;
                        aiMatchAnalysis: string | null;
                        tailoredResumeText: string | null;
                        tailoredResumePdfUrl: string | null;
                    }[];
                };
            };
            /** @description Response for status 500 */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        message: string;
                        code?: number;
                    };
                };
            };
        };
    };
    postApiV1Job: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    id?: number;
                    title: string;
                    url?: string | null;
                    description?: string | null;
                    status?: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                    nextEventDate?: Record<string, never> | null;
                    lastCommuncationDate?: Record<string, never> | null;
                    notes?: string | null;
                    aiSummary?: string | null;
                    aiMatchAnalysis?: string | null;
                    tailoredResumeText?: string | null;
                    tailoredResumePdfUrl?: string | null;
                };
                "application/x-www-form-urlencoded": {
                    id?: number;
                    title: string;
                    url?: string | null;
                    description?: string | null;
                    status?: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                    nextEventDate?: Record<string, never> | null;
                    lastCommuncationDate?: Record<string, never> | null;
                    notes?: string | null;
                    aiSummary?: string | null;
                    aiMatchAnalysis?: string | null;
                    tailoredResumeText?: string | null;
                    tailoredResumePdfUrl?: string | null;
                };
                "multipart/form-data": {
                    id?: number;
                    title: string;
                    url?: string | null;
                    description?: string | null;
                    status?: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                    nextEventDate?: Record<string, never> | null;
                    lastCommuncationDate?: Record<string, never> | null;
                    notes?: string | null;
                    aiSummary?: string | null;
                    aiMatchAnalysis?: string | null;
                    tailoredResumeText?: string | null;
                    tailoredResumePdfUrl?: string | null;
                };
            };
        };
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: number;
                        title: string;
                        url: string | null;
                        description: string | null;
                        status: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                        nextEventDate: Record<string, never> | null;
                        lastCommuncationDate: Record<string, never> | null;
                        notes: string | null;
                        aiSummary: string | null;
                        aiMatchAnalysis: string | null;
                        tailoredResumeText: string | null;
                        tailoredResumePdfUrl: string | null;
                    };
                };
            };
            /** @description Response for status 500 */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        message: string;
                        code?: number;
                    };
                };
            };
        };
    };
    putApiV1JobById: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: string;
            };
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    id?: number;
                    title?: string;
                    url?: string | null;
                    description?: string | null;
                    status?: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                    nextEventDate?: Record<string, never> | null;
                    lastCommuncationDate?: Record<string, never> | null;
                    notes?: string | null;
                    aiSummary?: string | null;
                    aiMatchAnalysis?: string | null;
                    tailoredResumeText?: string | null;
                    tailoredResumePdfUrl?: string | null;
                };
                "application/x-www-form-urlencoded": {
                    id?: number;
                    title?: string;
                    url?: string | null;
                    description?: string | null;
                    status?: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                    nextEventDate?: Record<string, never> | null;
                    lastCommuncationDate?: Record<string, never> | null;
                    notes?: string | null;
                    aiSummary?: string | null;
                    aiMatchAnalysis?: string | null;
                    tailoredResumeText?: string | null;
                    tailoredResumePdfUrl?: string | null;
                };
                "multipart/form-data": {
                    id?: number;
                    title?: string;
                    url?: string | null;
                    description?: string | null;
                    status?: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                    nextEventDate?: Record<string, never> | null;
                    lastCommuncationDate?: Record<string, never> | null;
                    notes?: string | null;
                    aiSummary?: string | null;
                    aiMatchAnalysis?: string | null;
                    tailoredResumeText?: string | null;
                    tailoredResumePdfUrl?: string | null;
                };
            };
        };
        responses: {
            /** @description Response for status 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: number;
                        title: string;
                        url: string | null;
                        description: string | null;
                        status: ("rejected" | "applied" | "interview hr" | "interview tech" | "interview final" | "test task" | "offer" | "accepted") | null;
                        nextEventDate: Record<string, never> | null;
                        lastCommuncationDate: Record<string, never> | null;
                        notes: string | null;
                        aiSummary: string | null;
                        aiMatchAnalysis: string | null;
                        tailoredResumeText: string | null;
                        tailoredResumePdfUrl: string | null;
                    };
                };
            };
            /** @description Response for status 404 */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        message: string;
                        code?: number;
                    };
                };
            };
            /** @description Response for status 500 */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        message: string;
                        code?: number;
                    };
                };
            };
        };
    };
// ...
}